pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "helpingbot"
version = "0.1.0"
description = "Modular trading bot reference implementation"
authors = [{name = "Helping Bot"}]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0",
]

[tool.setuptools]
packages = [
    "application",
    "config",
    "domain",
    "infrastructure",
    "presentation",
    "risk",
    "strategies",
    "utils",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "coverage[toml]>=7.3",
]

[tool.coverage.run]
branch = true
source = [
    "application",
    "config",
    "domain",
    "infrastructure",
    "presentation",
    "risk",
    "strategies",
    "utils",
]

[tool.coverage.report]
fail_under = 90
show_missing = true
skip_covered = true
